Refactoring and cleanup of content views for zipcontent (for H5P)
