+++ .gitignore
@@ -128,3 +128,6 @@ build_tools/crowdin-cli.jar
 
 # SQLite file
 SQLITE_MAX_VARIABLE_NUMBER.cache
+
+# don't ignore the TTF font for H5P
+!h5p-core-18.ttf
