+++ kolibri/core/content/templates/content/h5p.html
@@ -78,7 +78,7 @@
 
         </script>
 
-        <script type="text/javascript" src="dist/js/h5p-standalone-frame.js"></script>
+        <script type="text/javascript" src="dist/js/h5p-standalone-frame.min.js"></script>
 
         {% for path in jsfiles %}
         <script type="text/javascript" src="{{ path }}"></script>
